Devfeature - bring up to date new version

+++ StreamerBotLib/Overlay/Models/OverlayActionType.cs
@@ -1,8 +1,5 @@
 ﻿using StreamerBotLib.Overlay.Enums;
 
-using System;
-using System.Collections.Generic;
-
 namespace StreamerBotLib.Overlay.Models
 {
     /// <summary>
@@ -40,7 +37,7 @@ namespace StreamerBotLib.Overlay.Models
         /// <summary>
         /// The user provided length of the notification.
         /// </summary>
-        public int Duration { get; set; } = 0;
+        public decimal Duration { get; set; } = 0;
 
         /// <summary>
         /// The path to an image used in the event.
@@ -91,7 +88,7 @@ namespace StreamerBotLib.Overlay.Models
                 if (strings.Count > 6)
                 {
                     OverlayTypes type = (OverlayTypes)Enum.Parse(typeof(OverlayTypes), strings.Dequeue());
-                    int Duration = int.Parse(strings.Dequeue());
+                    decimal Duration = decimal.Parse(strings.Dequeue());
                     string User = strings.Dequeue();
                     string Msg = strings.Dequeue();
                     string action = strings.Dequeue();
